shopify-fastapi-insights/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
beautifulsoup4==4.12.2
google-generativeai==0.3.2
pydantic>=2.7.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography>=41.0.0,<42.0.0
asyncio-mqtt==0.13.0 